# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "API for managing customers, mechanics, inventory, and service tickets."
  version: "1.0.0"


schemes:
  - "https"

consumes:
  - "application/json"
produces:
  - "application/json"

securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Enter: Bearer <your_token_here>"

paths:
  /customers/login:
    post:
      tags:
        - Customers
      summary: "Login endpoint (generates token)"
      description: "Validates customer credentials and returns token."
      parameters:
        - in: body
          name: body
          description: "Email and password for login."
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              auth_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              message: "Login successful"
              status: "success"

  /customers:
    post:
      tags:
        - Customers
      summary: "Register new customer"
      description: "Creates a new customer account. Email must be unique."
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Customer successfully created"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"

    get:
      tags:
        - Customers
      summary: "Get all customers"
      security:
        - bearerAuth: []
      description: "Retrieves every customer in the system."
      responses:
        200:
          description: "Fetched customers successfully"
          schema:
            $ref: "#/definitions/AllCustomers"

  /customers/me:
    get:
      tags:
        - Customers
      summary: "Get logged in customer"
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Customer retrieved"
          schema:
            $ref: "#/definitions/CustomerResponse"

    put:
      tags:
        - Customers
      summary: "Update logged-in customer"
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Successfully updated customer"
          schema:
            $ref: "#/definitions/UpdateCustomerResponse"

    delete:
      tags:
        - Customers
      summary: "Delete logged-in customer"
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Customer successfully deleted"
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"

  /mechanics:
    post:
      tags:
        - Mechanics
      summary: "Create a mechanic"
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CreateMechanicPayload"
      responses:
        201:
          description: "Mechanic created"
          schema:
            $ref: "#/definitions/CreateMechanicResponse"

    get:
      tags:
        - Mechanics
      summary: "Get all mechanics"
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          type: integer
        - name: per_page
          in: query
          type: integer
      responses:
        200:
          description: "Mechanics retrieved"
          schema:
            $ref: "#/definitions/AllMechanics"

  /mechanics/{mechanic_id}:
    get:
      tags:
        - Mechanics
      summary: "Get mechanic by ID"
      security:
        - bearerAuth: []
      parameters:
        - name: mechanic_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Mechanic found"
          schema:
            $ref: "#/definitions/MechanicResponse"

    put:
      tags:
        - Mechanics
      summary: "Update mechanic"
      security:
        - bearerAuth: []
      parameters:
        - name: mechanic_id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UpdateMechanicPayload"
      responses:
        200:
          description: "Mechanic updated"
          schema:
            $ref: "#/definitions/UpdateMechanicResponse"

    delete:
      tags:
        - Mechanics
      summary: "Delete mechanic"
      security:
        - bearerAuth: []
      parameters:
        - name: mechanic_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Mechanic deleted"
          schema:
            $ref: "#/definitions/DeleteMechanicResponse"

  /mechanics/most-active:
    get:
      tags:
        - Mechanics
      summary: "Get most active mechanic"
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Most active mechanic(s)"
          schema:
            $ref: "#/definitions/AllMechanics"

  /inventory:
    post:
      tags:
        - Inventory
      summary: "Create inventory item"
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CreateInventoryPayload"
      responses:
        201:
          description: "Inventory item created"
          schema:
            $ref: "#/definitions/CreateInventoryResponse"

    get:
      tags:
        - Inventory
      summary: "Get all inventory items"
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          type: integer
        - name: per_page
          in: query
          type: integer
      responses:
        200:
          description: "Inventory retrieved"
          schema:
            $ref: "#/definitions/AllInventory"

  /inventory/{part_id}:
    get:
      tags:
        - Inventory
      summary: "Get inventory item"
      security:
        - bearerAuth: []
      parameters:
        - name: part_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Found part"
          schema:
            $ref: "#/definitions/InventoryResponse"

    put:
      tags:
        - Inventory
      summary: "Update inventory item"
      security:
        - bearerAuth: []
      parameters:
        - name: part_id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          schema:
            $ref: "#/definitions/UpdateInventoryPayload"
      responses:
        200:
          description: "Inventory updated"
          schema:
            $ref: "#/definitions/UpdateInventoryResponse"

    delete:
      tags:
        - Inventory
      summary: "Delete inventory item"
      security:
        - bearerAuth: []
      parameters:
        - name: part_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Part deleted"
          schema:
            $ref: "#/definitions/DeleteInventoryResponse"

  /service-tickets:
    post:
      tags:
        - Service Tickets
      summary: "Create a service ticket"
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CreateTicketPayload"
      responses:
        201:
          description: "Ticket created"
          schema:
            $ref: "#/definitions/TicketResponse"

    get:
      tags:
        - Service Tickets
      summary: "Get all service tickets"
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          type: integer
        - name: per_page
          in: query
          type: integer
      responses:
        200:
          description: "Tickets retrieved"
          schema:
            $ref: "#/definitions/AllTickets"

  /service-tickets/my-tickets:
    get:
      tags:
        - Service Tickets
      summary: "Get logged-in customer's tickets"
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Tickets retrieved"
          schema:
            $ref: "#/definitions/AllTickets"

  /service-tickets/{ticket_id}/assign-mechanic/{mechanic_id}:
    put:
      tags:
        - Service Tickets
      summary: "Assign mechanic to ticket"
      security:
        - bearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
        - name: mechanic_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Mechanic assigned"
          schema:
            $ref: "#/definitions/TicketResponse"

  /service-tickets/{ticket_id}/remove-mechanic/{mechanic_id}:
    put:
      tags:
        - Service Tickets
      summary: "Remove mechanic from ticket"
      security:
        - bearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
        - name: mechanic_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Mechanic removed"
          schema:
            $ref: "#/definitions/TicketResponse"

  /service-tickets/{ticket_id}/add-part/{part_id}:
    put:
      tags:
        - Service Tickets
      summary: "Add part to ticket"
      security:
        - bearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          type: integer
          required: true
        - name: part_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: "Part added"
          schema:
            $ref: "#/definitions/TicketResponse"

definitions:
  LoginCredentials:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
    required:
      - email
      - password

  LoginResponse:
    type: object
    properties:
      auth_token:
        type: string
      message:
        type: string
      status:
        type: string

  CreateCustomerPayload:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      password:
        type: string
    required:
      - name
      - email
      - password

  CreateCustomerResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      phone:
        type: string

  AllCustomers:
    type: array
    items:
      $ref: "#/definitions/CreateCustomerResponse"

  CustomerResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      phone:
        type: string

  UpdateCustomerPayload:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      password:
        type: string

  UpdateCustomerResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      phone:
        type: string

  DeleteCustomerResponse:
    type: object
    properties:
      message:
        type: string

  CreateMechanicPayload:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      salary:
        type: number
    required:
      - name
      - email
      - salary

  CreateMechanicResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      salary:
        type: number

  AllMechanics:
    type: array
    items:
      $ref: "#/definitions/CreateMechanicResponse"

  MechanicResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      salary:
        type: number

  UpdateMechanicPayload:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      salary:
        type: number

  UpdateMechanicResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      salary:
        type: number

  DeleteMechanicResponse:
    type: object
    properties:
      message:
        type: string

  CreateInventoryPayload:
    type: object
    properties:
      name:
        type: string
      price:
        type: number
    required:
      - name
      - price

  CreateInventoryResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: number

  AllInventory:
    type: array
    items:
      $ref: "#/definitions/CreateInventoryResponse"

  InventoryResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: number

  UpdateInventoryPayload:
    type: object
    properties:
      name:
        type: string
      price:
        type: number

  UpdateInventoryResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: number

  DeleteInventoryResponse:
    type: object
    properties:
      message:
        type: string

  CreateTicketPayload:
    type: object
    properties:
      VIN:
        type: string
      service_date:
        type: string
      service_desc:
        type: string
    required:
      - VIN
      - service_desc

  TicketResponse:
    type: object
    properties:
      id:
        type: integer
      VIN:
        type: string
      service_date:
        type: string
      service_desc:
        type: string
      customer_id:
        type: integer

  AllTickets:
    type: array
    items:
      $ref: "#/definitions/TicketResponse"